refactor: Score, game 패키지로 이동
